backend/requirements.txt
# MongoDB Atlas Vector Search Setup Dependencies
pymongo==4.6.0
python-dotenv==1.0.0

# Optional: For advanced MongoDB operations
motor==3.3.2  # Async MongoDB driver
dnspython==2.4.2  # For MongoDB SRV connection strings

# Optional: For data processing and embeddings
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Optional: For API development (if building backend endpoints)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Optional: For testing
pytest==7.4.3
pytest-asyncio==0.21.1
